# compiler: create eval vm lazily